+++ app/src/main/java/com/android/shaftschematic/util/HintStyle.kt
@@ -1,3 +1,0 @@
-package com.android.shaftschematic.util
-
-enum class HintStyle { CHIP, TEXT, OFF }